(react app) make game overview a component
